+++ examples/custom_middleware.rs
@@ -1,8 +1,8 @@
 // Custom middleware example.
 // All bytes that go through the protocol are rotated by an offset of 13.
 
-#[macro_use] extern crate protocol;
-#[macro_use] extern crate protocol_derive;
+#[macro_use] extern crate djin_protocol;
+#[macro_use] extern crate djin_protocol_derive;
 
 use std::num::Wrapping;
 
@@ -21,13 +21,13 @@ impl RotateMiddleware
     }
 }
 
-impl protocol::wire::Middleware for RotateMiddleware
+impl djin_protocol::wire::Middleware for RotateMiddleware
 {
-    fn decode_data(&mut self, data: Vec<u8>) -> Result<Vec<u8>, protocol::Error> {
+    fn decode_data(&mut self, data: Vec<u8>) -> Result<Vec<u8>, djin_protocol::Error> {
         Ok(data.into_iter().map(|byte| (Wrapping(byte) - Wrapping(self.offset)).0).collect())
     }
 
-    fn encode_data(&mut self, data: Vec<u8>) -> Result<Vec<u8>, protocol::Error> {
+    fn encode_data(&mut self, data: Vec<u8>) -> Result<Vec<u8>, djin_protocol::Error> {
         Ok(data.into_iter().map(|byte| (Wrapping(byte) + Wrapping(self.offset)).0).collect())
     }
 }
@@ -62,7 +62,7 @@ fn main() {
     use std::net::TcpStream;
 
     let stream = TcpStream::connect("127.0.0.1:34254").unwrap();
-    let mut connection = protocol::wire::stream::Connection::new(stream, Pipeline::new(), protocol::Settings::default());
+    let mut connection = djin_protocol::wire::stream::Connection::new(stream, Pipeline::new(), djin_protocol::Settings::default());
 
     connection.send_packet(&Packet::Ping(Ping { id: 0, data: vec![ 55 ]})).unwrap();