+++ protocol-derive/src/codegen/mod.rs
@@ -49,14 +49,14 @@ pub fn named_fields_declarations(fields: &syn::Fields) -> TokenStream {
                 quote! {
                     #update_hints_fixed
                     __hints.set_skip(#skip_condition);
-                    let #field_name: #field_ty = protocol::Parcel::read_field(__io_reader, __settings, &mut __hints)?;
+                    let #field_name: #field_ty = djin_protocol::Parcel::read_field(__io_reader, __settings, &mut __hints)?;
                     #update_hints
                     __hints.next_field();
                 }
             } else {
                 quote! {
                     #update_hints_fixed
-                    let #field_name: #field_ty = protocol::Parcel::read_field(__io_reader, __settings, &mut __hints)?;
+                    let #field_name: #field_ty = djin_protocol::Parcel::read_field(__io_reader, __settings, &mut __hints)?;
                     #update_hints
                     __hints.next_field();
                 }
@@ -91,7 +91,7 @@ fn read_named_fields_enum(fields_named: &syn::FieldsNamed)
         quote! {
             #field_name : {
                 #update_hints_fixed
-                let res: Result<#field_ty, _> = protocol::Parcel::read_field(__io_reader, __settings, &mut __hints);
+                let res: Result<#field_ty, _> = djin_protocol::Parcel::read_field(__io_reader, __settings, &mut __hints);
                 #update_hints
                 __hints.next_field();
                 res?
@@ -159,7 +159,7 @@ fn update_hint_fixed_length<'a>(field: &'a syn::Field,
         let position = fields.clone().into_iter().position(|f| f == field).unwrap();
 
         quote! {
-            __hints.set_field_length(#position, #length, protocol::hint::LengthPrefixKind::Elements);
+            __hints.set_field_length(#position, #length, djin_protocol::hint::LengthPrefixKind::Elements);
         }
     } else {
         quote! { }
@@ -259,7 +259,7 @@ fn write_named_fields(fields_named: &syn::FieldsNamed)
         quote! {
             {
                 #update_hints_fixed
-                let res = protocol::Parcel::write_field(&self. #field_name, __io_writer, __settings, &mut __hints);
+                let res = djin_protocol::Parcel::write_field(&self. #field_name, __io_writer, __settings, &mut __hints);
                 #update_hints
                 __hints.next_field();
                 res?
@@ -277,7 +277,7 @@ fn read_unnamed_fields(fields_unnamed: &syn::FieldsUnnamed)
 
         quote! {
             {
-                let res: Result<#field_ty, _> = protocol::Parcel::read_field(__io_reader, __settings, &mut __hints);
+                let res: Result<#field_ty, _> = djin_protocol::Parcel::read_field(__io_reader, __settings, &mut __hints);
                 __hints.next_field();
                 res?
             }
@@ -294,7 +294,7 @@ fn write_unnamed_fields(fields_unnamed: &syn::FieldsUnnamed)
     let field_writers: Vec<_> = field_indices.map(|field_index| {
         quote! {
             {
-                let res = protocol::Parcel::write_field(&self. #field_index, __io_writer, __settings, &mut __hints);
+                let res = djin_protocol::Parcel::write_field(&self. #field_index, __io_writer, __settings, &mut __hints);
                 __hints.next_field();
                 res?
             }