+++ tests/src/hints/enums.rs
@@ -1,4 +1,4 @@
-use protocol::{hint, Settings};
+use djin_protocol::{hint, Settings};
 use super::{SaveHints, HasSavedHints};
 
 #[derive(Protocol, Debug, PartialEq)]
@@ -29,7 +29,7 @@ define_common_hint_invariant_tests!(with_named_fields => WithNamedFields : WithN
 define_common_hint_invariant_tests!(with_unnamed_fields => WithUnnamedFields : WithUnnamedFields::default());
 
 mod named_fields {
-    use protocol::Parcel;
+    use djin_protocol::Parcel;
     use super::*;
 
     #[test]
@@ -53,7 +53,7 @@ mod named_fields {
 }
 
 mod unnamed_fields {
-    use protocol::Parcel;
+    use djin_protocol::Parcel;
     use super::*;
 
     #[test]