+++ tests/src/logic/aligned.rs
@@ -1,4 +1,4 @@
-use protocol::{Parcel, Settings};
+use djin_protocol::{Parcel, Settings};
 
 /// An example packet with a length prefix disjoint
 /// from its data, with the data also
@@ -9,7 +9,7 @@ struct Packet {
     /// The version number of the protocol.
     pub version_number: (u32, u32),
     #[protocol(length_prefix(bytes(reason_length)))]
-    pub reason: protocol::logic::Aligned<String, u64>,
+    pub reason: djin_protocol::logic::Aligned<String, u64>,
 }
 
 #[test]
@@ -18,7 +18,7 @@ fn write_alignment_pads_zero() {
         reason_length: 12,
         version_number: (11, 0xdeadbeef),
         reason: "hello world!".to_owned().into(),
-    }.raw_bytes(&protocol::Settings::default()).unwrap();
+    }.raw_bytes(&djin_protocol::Settings::default()).unwrap();
     assert_eq!(&[
         12, // reason length
         0, 0, 0, 11, 0xde, 0xad, 0xbe, 0xef, // version number